printer/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0